contrib/gdk-pixbuf-xlib: Fix rgb888amsb()
On Little Endian, it wrote memory out of bounds (i.e. when running
inside a little endian client, connecting to a big endian X server).

On Big Endian, it only processed the first 1/4 of the icon.

The effects can be seen when running IceWM. Running native on Big Endian
ppc64 with a Big Endian X11 server, the icon background was 3/4 random:

  https://bugzilla.suse.com/show_bug.cgi?id=929462

This match makes rgb888amsb() behave analogously to rgb888msb().

Signed-off-by: Max Staudt <mstaudt@suse.de>