io-bmp: Add overflow checks to BMP image saver
Return an error if the image width or height are too big to fit in the
BMP size fields.

This bumps our GLib dependency to 2.48.0, as it uses
g_uint_checked_mul(). If you want to backport this fix, it’s probably
best to use __builtin_umul_overflow() directly (if compiling using GCC
or another compiler which supports it).

Coverity ID: 1388532

https://bugzilla.gnome.org/show_bug.cgi?id=777315